pyproject.toml
[tool.poetry]
name = "kenzie-commerce-grupo-29-t14"
version = "0.1.0"
description = "Final project M5"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"
drf-spectacular = "^0.26.2"
django-money = "^3.1.0"
django-cors-headers = "^3.14.0"
dj-database-url = "^2.0.0"
whitenoise = {extras = ["brotli"], version = "^6.4.0"}
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
black = "^23.3.0"
ipdb = "^0.13.13"
pycodestyle = "^2.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
